# MAINT: speed up hstack and vstack by eliminating list comprehension.