While discussing optimizations in #13634, it was clear that there was a simple optimization that was missed: `atleast_*d` was called on every array rather than just passing it the whole tuple (causing dispatch overhead for each if nothing else). Removing that,

```
x = np.array([1])
xs = [x, x, x]
%timeit np.hstack(xs)
# 4.33 ->3.45 us
%timeit np.vstack(xs)
# 5.35->4.33 us
```