@mhvk 2019-06-02
@seberg - the request to you was just because you were the "suggested" reviewer and I clicked wrong. 

@shoyer - since this was discussed in #13634, I thought you might do a quick review.