@mhvk 2019-06-02
OK, I added `dstack` - but didn't move it in this PR (probably best done as a more general merge - quite topical, though, given discussion on mailing list!)