@mhvk 2019-06-02
Yes, I even had `arrs = atleast_1d(tup) if len(tup) > 1 else [atleast_1d(tup)]` at some point but ended disliking the two separate invocations.

Anyway, will merge since both of you approved.